LPS-55590 Copy EditOrgLaborAction, no logic changes
